fix: InviteCode 갱신 메서드 이동
